Enable distributed computations for logloss_function primitive, implement tests
